Move loading of command into another file, add coded description
