+++ src/app.ts
@@ -5,26 +5,14 @@ import { Telegraf } from "telegraf";
 import callbackEvent from "./events/callback";
 import readyEvent from "./events/ready";
 import { createConnection } from "typeorm";
-import { promises as fs } from "fs";
-import { resolve } from "path";
+import { loadCommands } from "./helpers/loadCommands";
 
 export const bot = new Telegraf(process.env.TOKEN);
 
 bot.on("callback_query", callbackEvent);
 
 async function bootstrap() {
-    const commandsDirPath = resolve(__dirname, "commands");
-    const cmds = (await fs.readdir(commandsDirPath, { withFileTypes: true }))
-        .map((f) => f.name)
-        .filter((name) => !name.includes("index"));
-
-    for (const command of cmds) {
-        const file = (await import(resolve(commandsDirPath, command))).default;
-        const commandName = command.split(".")[0];
-        bot.command(commandName, file);
-        console.info(`Command ${commandName} loaded`);
-    }
-
+    await loadCommands();
     await createConnection();
     await bot.launch();
     await readyEvent();
